@@ -1568,13 +1568,7 @@ class RenderWindowInteractor(_NoNewAttrMixin):
     def terminate_app(self):
         """Terminate the app."""
         if self.initialized:
-            # #################################################################
-            # 9.0.2+ compatibility:
-            # See: https://gitlab.kitware.com/vtk/vtk/-/issues/18242
-            if hasattr(self.interactor, 'GetDone'):
-                self.interactor.SetDone(True)
-            # #################################################################
-
+            self.interactor.SetDone(True)  # See: https://gitlab.kitware.com/vtk/vtk/-/issues/18242
             self.interactor.TerminateApp()
 
     def close(self):
